backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
groq==0.4.1
python-dotenv==1.0.0
pydantic==2.5.0
pandas==2.1.4
numpy==1.24.3
python-multipart==0.0.6
cors==1.0.1
fastapi-cors==0.0.6
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
sqlalchemy==2.0.23
Jinja2==3.1.2